% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_filter_dmps.R
\name{apply_filter_dmps}
\alias{apply_filter_dmps}
\title{Apply Filtered DMPs and Generate Plots}
\usage{
apply_filter_dmps(
  dmps,
  dev = "png",
  p.value = seq(0, 0.1, 0.01),
  mDiff = seq(0.15, 0.5, 0.05),
  path = "analysis/intermediate/dmps/"
)
}
\arguments{
\item{dmps}{Data table containing DMPs.}

\item{dev}{Device to use for plotting (e.g., "png").}

\item{p.value}{Vector of p-values to explore.}

\item{mDiff}{Vector of mean methylation differences to explore.}

\item{path}{Path to save the plots and intermediate files.}
}
\value{
None (Plots are saved to the specified path).
}
\description{
Apply Filtered DMPs and Generate Plots
}
