% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_filter_dmrs.R
\name{apply_filter_dmrs}
\alias{apply_filter_dmrs}
\title{Apply Filtering to DMRs}
\usage{
apply_filter_dmrs(
  dmrs,
  plots = TRUE,
  p.value = seq(0.001, 0.11, 0.025),
  mDiff = seq(0.15, 0.5, 0.05),
  min.cpg = seq(3, 6, 1),
  path = "analysis/intermediate/dmrs/"
)
}
\arguments{
\item{dmrs}{Data.table containing DMR information.}

\item{plots}{Logical, indicating whether to generate plots.}

\item{p.value}{Vector of p-values.}

\item{mDiff}{Vector of methylation differences.}

\item{min.cpg}{Vector of minimum CpG values.}

\item{path}{Path to save intermediate DMR files.}
}
\value{
A melted data.table containing the filtered DMRs.
}
\description{
This function applies filtering to Differentially Methylated Regions (DMRs) and generates plots.
}
\examples{
# Example usage:
filtered_dmrs <- apply_filter_dmrs(dmrs = dmrs_data, plots = TRUE)

}
\author{
Izar de Villasante
}
