% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betas_disk.R
\name{betas_disk}
\alias{betas_disk}
\title{Use Betas Object in Parallel and Save to Disk}
\usage{
betas_disk(rgSet)
}
\arguments{
\item{rgSet}{A SummarizedExperiment object containing beta values.}
}
\value{
A bigstatsr::FBM object representing the beta matrix.
}
\description{
This function takes a beta matrix from a rgSet object and converts it into a
bigstatsr::FBM (Filebacked Big Matrix) for parallel processing. This can help
in efficient storage and manipulation of beta values, especially when the
matrix is large. The betas are saved to disk using the bigstatsr package.
}
