% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bplot.R
\name{bplot}
\alias{bplot}
\title{Generate PCA bi-plots}
\usage{
bplot(
  pca,
  ss,
  colgroup,
  s,
  combs = NULL,
  pal = NULL,
  tit = NULL,
  labs = TRUE,
  overlap = Inf,
  alfa = 0.3,
  folder = "analysis/pca/bplots/",
  idcol = "Sample_Name"
)
}
\arguments{
\item{pca}{A prcomp object (output of PCA analysis).}

\item{ss}{Data frame or data table containing samplesheet with metadata information.}

\item{colgroup}{Character column name in ss. Color points based on this variable.}

\item{s}{Character column name in ss. Shape points according to categories of that variable.}

\item{combs}{2D matrix (default = combn(4,2)) mapping PCs to plot (1st row = X, 2nd row = Y).}

\item{tit}{Title for the plot.}

\item{labs}{Logical. Add labels if TRUE (default).}

\item{overlap}{Numeric (default = Inf). Pass to max.overlaps in geom_text_repel.}

\item{alfa}{Alpha value for transparency (default = 0.3).}

\item{folder}{Path to folder where plots are saved (default: "analysis/pca/bplots/").}

\item{idcol}{Column name for the sample ID (default: "Sample_Name").}

\item{cols}{Color palette to use (default: get_cols() function).}
}
\value{
Path to the folder where results are saved.
}
\description{
Generate PCA bi-plots
}
\details{
This function generates PCA bi-plots using ggplot2 and related packages.
}
\author{
izar de Villasante
}
