% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution.R
\name{celldeconvolution}
\alias{celldeconvolution}
\title{Cell type Deconvolution}
\usage{
celldeconvolution(raw_object, object, arraytype)
}
\arguments{
\item{raw_object}{The input raw RGChannelSet.}

\item{object}{MehtylSet object with normalized data.}

\item{arraytype}{Vector containing the name of the array platform. Options: '450K', 'EPIC', 'EPICv2'.}
}
\value{
A MethylSet object containing deconvolution results in the colData slot.
}
\description{
This function performs deconvolution based on DNA methylation data, estimating cell type proportions
using raw and normalized methylation datasets. It supports both HM450 and EPIC platforms.
}
\examples{
data(rgSet)
data("mSet_normalized")
deconvoluted<-celldeconvolution(rgSet,mSet_normalized,arraytype='EPICv2')
colData(deconvoluted)
}
