% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_list.R
\name{correlated_list}
\alias{correlated_list}
\title{List of correlated variables}
\usage{
correlated_list(dataset)
}
\arguments{
\item{dataset}{Sample sheet}
}
\value{
A list with correlated variables
}
\description{
Obtain a list of variables that are correlated or its design is not well balanced, depending on each type of variables(numerical/categorical)
}
