% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_list.R
\name{correlation_analysis}
\alias{correlation_analysis}
\title{Correlation analysis}
\usage{
correlation_analysis(clean_object, path)
}
\arguments{
\item{clean_object}{Rgset object}

\item{path}{Path to save the correlation plot}
}
\value{
List of correlated variables and a plot
}
\description{
This function performs correlation analysis between metadata variables.
}
