% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations_plot.R
\name{correlation_plot}
\alias{correlation_plot}
\title{Correlation analysis plot}
\usage{
correlation_plot(correlation_list, dataset, sg, path = "./")
}
\arguments{
\item{correlation_list}{A correlation list containing the names of the variables that are correlated}

\item{dataset}{Sample sheet}

\item{path}{Path to save the plot (default: './')}
}
\description{
Pairwise correlation plots. Plots are colored if the variables are correlated.
}
