% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_datatable.R
\name{custom_datatable}
\alias{custom_datatable}
\title{Create DataTable with custom options.}
\usage{
custom_datatable(data)
}
\arguments{
\item{data}{A data frame to be displayed in the DataTable.}
}
\value{
A DataTable with custom options.
}
\description{
This function creates a DataTable with custom options, including download buttons for the entire dataset.
}
\examples{
\dontrun{
dtable(my_data)
}
}
\keyword{internal}
