% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_list.R
\name{dataset_var}
\alias{dataset_var}
\title{Select Variables from the Dataset}
\usage{
dataset_var(dataset, variables = "ALL")
}
\arguments{
\item{dataset}{A data.table or data.frame containing the dataset.}

\item{variables}{A vector of variable names to select. If set to \code{'ALL'}, the full dataset is returned.}
}
\value{
A subset of the dataset containing only the specified variables, or the full dataset if \code{'ALL'} is specified.
}
\description{
This function selects specified variables from a dataset. If no variables are specified, the full dataset is returned.
}
