% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denplot.R
\name{denplot}
\alias{denplot}
\title{Density plot after normalization}
\usage{
denplot(object, metadata, sampGroups, path = "./")
}
\arguments{
\item{object}{A MethylSet object with normalized data}

\item{metadata}{Sample sheet information}

\item{sampGroups}{Variable of the metadata object for coloring groups}

\item{path}{path to the folder where plots will be saved. (Default ./)}
}
\value{
plots
}
\description{
This function generates a density plot of beta-values distribution after normalization.
}
