% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC_plots.R
\name{densPlot}
\alias{densPlot}
\title{Density Plot for Beta Values (ggplot2)}
\usage{
densPlot(
  dat,
  sg = NULL,
  main = "",
  xlab = "Beta",
  pal = RColorBrewer::brewer.pal(8, "Dark2"),
  cols = NULL,
  xlim,
  ylim,
  add = TRUE,
  legend = TRUE,
  ...
)
}
\arguments{
\item{dat}{Data object. Must be an 'RGChannelSet', a 'MethylSet', or a matrix.}

\item{sg}{Sample groups. If NULL, defaults to one group for all samples.}

\item{main}{Main title for the plot.}

\item{xlab}{Label for the x-axis.}

\item{pal}{Color palette for sample groups (default: Brewer palette "Dark2").}

\item{cols}{Vector of colors for sample groups. If NULL, colors are generated using get_cols.}

\item{xlim}{Range for the x-axis.}

\item{ylim}{Range for the y-axis.}

\item{add}{Logical, whether to add to an existing plot.}

\item{legend}{Logical, whether to include a legend.}
}
\value{
A ggplot object.
}
\description{
Generates a density plot for beta values from an 'RGChannelSet', 'MethylSet', or matrix using ggplot2.
}
