% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_plots.R
\name{distribution_plots}
\alias{distribution_plots}
\title{ttest_boxplot}
\usage{
distribution_plots(data, variable_interest, path = "./")
}
\arguments{
\item{data}{Data frame containing metadata}

\item{variable_interest}{Name of the variable in the metadata that refers to the sample Grops or condition.}

\item{path}{Path to save the plots (default: "./")}
}
\description{
Distribution plots
}
\details{
Create boxplots or barplots for all the variables, colored by the variable of interest
}
