% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMPextr.R
\name{DMPextr}
\alias{DMPextr}
\title{Extract DMPs}
\usage{
DMPextr(
  fit,
  ContrastsDM = colnames(fit$contrasts),
  p.value,
  beta_normalized,
  betas_idx = NULL,
  mDiff,
  ann = NULL,
  writedir = "analysis/DMP_",
  writeOut = TRUE,
  ncores = NULL,
  columns = TRUE
)
}
\arguments{
\item{fit}{A \code{limma} fit object (e.g., result from \code{lmFit()} or \code{eBayes()}).}

\item{ContrastsDM}{A list of contrasts as returned by \code{limma::makeContrasts()}.}

\item{p.value}{P-value threshold for filtering DMPs based on adjusted p-value.}

\item{beta_normalized}{A matrix of normalized beta values.}

\item{betas_idx}{An index for subsetting beta values.}

\item{mDiff}{Absolute mean methylation difference between groups to filter by.}

\item{ann}{Annotation dataset from manifest with metadata such as gene information, CGI, RefGene, etc.}

\item{writedir}{Directory for saving output files.}

\item{writeOut}{Logical, whether to save results as CSV files (default: \code{TRUE}).}

\item{ncores}{Number of cores to use for parallel processing (optional).}

\item{columns}{Optional columns to select from \code{limma::topTable}.}
}
\value{
A \code{data.table} containing the extracted DMPs, including annotation, methylation values, and statistical information.
}
\description{
This function processes a set of contrasts, extracts Differentially Methylated Positions (DMPs) using \code{limma::topTable},
and adds gene annotation and methylation values. The function also filters DMPs based on mean methylation difference
and allows for saving the results to CSV files.
}
