% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{Dtable}
\alias{Dtable}
\title{Shiny App with DataTable and Download Buttons.}
\arguments{
\item{data}{A data frame to be displayed in the DataTable.}
}
\value{
A Shiny app with a DataTable and download buttons.
}
\description{
This Shiny app creates a DataTable with custom options, including download buttons for the entire dataset.
}
\examples{
\dontrun{
dtable(my_data)
}
}
