% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_methods.R
\name{Em}
\alias{Em}
\title{Apply ENmix normalization to RGChannelSet and map to genome.}
\usage{
Em(rgSet, arraytype = NULL)
}
\arguments{
\item{rgSet}{RGChannelSet object to be normalized.}

\item{arraytype}{Type of array used for preprocessing. If provided, array-specific annotation is used.}
}
\value{
Normalized GenomicRatioSet object.
}
\description{
Apply ENmix normalization to RGChannelSet and map to genome.
}
