% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter}
\alias{filter}
\title{Filter RGChannelSet based on Quality Control Criteria}
\usage{
filter(
  rgSet,
  sampGroups = NULL,
  sampNames = "Sample_Name",
  save_barplot = TRUE,
  frac = 0.1,
  pval = 0.01,
  arraytype = NULL,
  cols = NULL,
  qc_folder = "analysis/intermediate/QC"
)
}
\arguments{
\item{rgSet}{RGChannelSet object to be filtered.}

\item{sampGroups}{Column name indicating sample groups.}

\item{sampNames}{Column name in 'rgSet@colData' containing sample names.}

\item{save_barplot}{Logical, whether to save the barplot of mean detection p-values.}

\item{frac}{Fraction of probes not passing the p-value filter for sample exclusion.}

\item{pval}{P-value threshold for sample and probe filtering.}

\item{arraytype}{Array type, used to update annotation for Illumina arrays.}

\item{cols}{Vector of colors for sample groups. If NULL, colors are generated using get_cols.}

\item{qc_folder}{Folder path for saving quality control plots.}
}
\value{
Filtered RGChannelSet.
}
\description{
This function filters an RGChannelSet based on quality control criteria.
It removes low-quality samples and probes, and generates a barplot of mean detection p-values.
}
\examples{

data("rgSet")

# Filter RGChannelSet based on quality control criteria
rgSet_filtered<-filter(rgSet, sampGroups = 'Type', sampNames = "Sample_Name",qc_folder = './')

}
