% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_filter_dmrs.R
\name{filter_dmrs}
\alias{filter_dmrs}
\title{Filter DMRs}
\usage{
filter_dmrs(dmrs, p.value = 0.01, mDiff = 0.2, min.cpg = 5, s = FALSE)
}
\arguments{
\item{dmrs}{Data.table containing DMR information.}

\item{p.value}{False Discovery Rate threshold.}

\item{mDiff}{Methylation difference threshold.}

\item{min.cpg}{Minimum number of CpGs in a DMR.}

\item{s}{Logical, indicating whether to generate a summary of filtered DMRs.}
}
\value{
A data.table containing the filtered DMRs.
}
\description{
This function filters Differentially Methylated Regions (DMRs) based on specified criteria.
}
\author{
Izar de Villasante
}
