% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_methods.R
\name{funn}
\alias{funn}
\title{Apply FunNorm normalization to RGChannelSet Object}
\usage{
funn(rgSet)
}
\arguments{
\item{rgSet}{RGChannelSet object to normalize.}
}
\value{
Normalized RGChannelSet object.
}
\description{
This function applies the FunNorm (Functional Normalization) method to an RGChannelSet object. It corrects for biases due to the probe design or other unwanted systematic variations.
}
