% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_category.R
\name{get_category}
\alias{get_category}
\title{Get Category from Input Sample Sheet}
\usage{
get_category(object)
}
\arguments{
\item{object}{data.table or data.frame, sample sheet.}
}
\value{
A vector with each column's category.
}
\description{
Sets the category for each variable in the sample sheet so information in each column is treated properly ("ids", "covs", "batch", "mgroup").
}
