% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_arraytype.R
\name{guess_arraytype}
\alias{guess_arraytype}
\title{Guess type of DNA methylation array.}
\usage{
guess_arraytype(n_probes)
}
\arguments{
\item{n_probes}{Number of probes detected}
}
\value{
Array type, one of:
\itemize{
\item IlluminaHumanMethylationEPIC
\item IlluminaHumanMethylation450k
\item IlluminaHumanMethylation27k
\item HorvathMammalMethylChip40
\item unknown
}
}
\description{
Will report the platform of a DNA methylation array given the number
of probes measured. Adapted from an internal function in \code{minfi} package.
Returns \emph{unknown} if array platform cannot be determined.
}
\keyword{internal}
