% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_sample_sheet.R
\name{input_sample_sheet}
\alias{input_sample_sheet}
\title{Load data target factory.}
\usage{
input_sample_sheet(file)
}
\arguments{
\item{file}{Character, Sample sheet file path.}
}
\value{
A list of target objects.
}
\description{
Define 4 targets:
\enumerate{
\item Track the user-supplied data file.
\item Read the data using \code{read_file()} .
\item Generate category to handle info in sample_sheet.
}
}
