% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_results_dirs.R
\name{make_results_dirs}
\alias{make_results_dirs}
\title{Create Directories for Data Storage in the Pipeline}
\usage{
make_results_dirs(
  subf,
  results_folder = "./results/",
  analysis_folder = "./analysis/"
)
}
\arguments{
\item{subf}{A subfolder name to organize results further.}

\item{results_folder}{The main folder to store all results. Default is "./results/".}

\item{analysis_folder}{The folder to store analysis-specific results. Default is "./analysis/".}
}
\value{
A list of folder paths created for different purposes.
}
\description{
This function creates directories for storing various types of results and plots
generated during the pipeline analysis.
}
\examples{
# Create directories with default paths
make_results_dirs(subf = "sample_analysis")

# Create directories with custom paths
make_results_dirs(results_folder = "./my_results/", analysis_folder = "./my_analysis/", subf = "custom_analysis")

}
