% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_filter_dmrs.R
\name{make_ribbon_names}
\alias{make_ribbon_names}
\title{Make Ribbon Names}
\usage{
make_ribbon_names(var)
}
\arguments{
\item{var}{Variable for which ribbon names are generated.}
}
\value{
A vector of ribbon names.
}
\description{
This function generates ribbon names based on a variable.
}
\examples{
# Example usage:
ribbon_names <- make_ribbon_names(var = "min.cpg")

}
\author{
Izar de Villasante
}
