% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanoplot.R
\name{manhattanplot}
\alias{manhattanplot}
\title{Generate Manhattan Plots from Genomic Data}
\usage{
manhattanplot(object, path = "./")
}
\arguments{
\item{object}{Data frame containing DMPs or DMRs.}

\item{path}{A character string specifying the directory to save the plot files. Default: current directory ("./")}
}
\description{
This function creates Manhattan plots for genomic data, saving the plots as PNG files.
The function uses \code{karyoploteR} to create the plots and highlights the top CpG sites with a suggestive threshold.
}
