% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.R
\name{mod}
\alias{mod}
\title{Construct models and contrasts with limma}
\usage{
mod(
  object,
  betas_idx = NULL,
  group_var = "Sample_Group",
  covs.formula = NULL,
  contrasts = NULL,
  covs = NULL,
  metadata,
  set = TRUE,
  gr = NULL,
  pairwise = TRUE,
  singular = FALSE,
  rename = NULL,
  idcol = "barcode"
)
}
\arguments{
\item{object}{Your object containing beta values.}

\item{betas_idx}{Index for beta values.}

\item{group_var}{The variable used as an independent variable.}

\item{covs.formula}{Formula for specifying covariates in the model.}

\item{contrasts}{Optional string of semicolon-separated contrasts. If not provided, automatic contrasts are generated.}

\item{covs}{A character vector specifying additional covariates.}

\item{metadata}{The metadata or sample sheet.}

\item{set}{A boolean vector to subset the observations.}

\item{gr}{A character vector specifying group variables.}

\item{pairwise}{Logical indicating whether to create pairwise group comparisons (default: TRUE).}

\item{singular}{Logical indicating whether to create singular group comparisons (default: FALSE).}

\item{rename}{Character vector to rename the contrasts.}

\item{idcol}{Column name for the sample ID (default: "barcode").}
}
\value{
A limma eBayes model.
}
\description{
Construct Models and Contrasts with limma
}
\details{
This function is used to construct models and contrasts for differential expression analysis using limma.
}
\author{
izar de Villasante
}
