% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myModal.R
\name{myModal}
\alias{myModal}
\title{Modal dialog with download buttons for internal use.}
\usage{
myModal()
}
\value{
A modal dialog with download buttons.
}
\description{
This function creates a modal dialog with download buttons for CSV and Excel files.
It is intended for internal use within the package and should not be exported to the end user.
}
\examples{
\dontrun{
myModal()
}
}
\keyword{internal}
