% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_methods.R
\name{noob_pq}
\alias{noob_pq}
\title{Apply Noob followed by Quantile normalization to RGChannelSet Object}
\usage{
noob_pq(rgSet)
}
\arguments{
\item{rgSet}{RGChannelSet object to normalize.}
}
\value{
Normalized RGChannelSet object.
}
\description{
Apply Noob followed by Quantile normalization to RGChannelSet Object
}
