% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_methods.R
\name{normalization_all_methods}
\alias{normalization_all_methods}
\title{Function to normalize data by all methods and plot them}
\usage{
normalization_all_methods(
  object,
  metadata,
  sampGroups,
  path = "./",
  norm_method
)
}
\arguments{
\item{object}{filtered object to be normalized}

\item{metadata}{Sample sheet}

\item{sampGroups}{Sample sheet name variable to color plots}

\item{path}{Path to save the plot}

\item{norm_method}{Normalization method that want to exlcude}
}
\description{
Function to normalize data by all methods and plot them
}
