% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway.R
\name{path_results}
\alias{path_results}
\title{Generate Results Excel Sheet for Pathway Analysis}
\usage{
path_results(
  pathway,
  topN = 50,
  method = "method",
  pval = 0.05,
  path = "results/pathways.csv",
  cols = NULL,
  savefile = FALSE
)
}
\arguments{
\item{pathway}{A data.table with pathway analysis results compatible with gopath function output values.}

\item{topN}{Numeric value. The minimum number of terms to include for each group.}

\item{method}{A character or a vector of characters specifying the columns to group by. Default is "method".}

\item{pval}{FDR threshold to filter out terms (default: 0.05).}

\item{path}{The path to save the results CSV file (default: "results/pathways.csv").}

\item{cols}{Columns to include in the pathway results (default: NULL).}

\item{savefile}{Logical indicating whether to save the results to a file (default: FALSE).}
}
\value{
A data.table containing filtered pathway analysis results.

A data.table containing filtered pathway analysis results.
}
\description{
Generate Pathway Results in an Excel Sheet
}
\details{
This function filters pathway analysis results for at least the topN pathways for each group and includes all terms with FDR <= FDR. The filtered results are then saved to a CSV file and returned as a data.table.
}
\seealso{
\code{\link{pathway}} for pathway analysis.
}
\keyword{analysis}
\keyword{data.table}
\keyword{filtering}
\keyword{pathway}
\keyword{results}
