% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway.R
\name{pathway}
\alias{pathway}
\title{Perform Pathway Enrichment Analysis for Differentially Methylated Positions or Regions (DMPs or DMRs)}
\usage{
pathway(
  dmrs,
  path = "results/pathways.csv",
  cols = c("term_size", "query_size", "intersection_size"),
  pval = 0.05,
  topN = 50,
  savefile = FALSE,
  re.folder = "./",
  enrich_plots = FALSE
)
}
\arguments{
\item{dmrs}{A data.table containing Differentially Methylated Positions or Regions (DMPs or DMRs), including at least the 'Gene' and 'Contrast' columns.}

\item{path}{Path to save the combined pathway analysis results (default: "results/pathways.csv").}

\item{cols}{Columns to include in the pathway results (default: c("term_size", "query_size", "intersection_size")).}

\item{pval}{P-value threshold for significance in pathway analysis (default: 0.05).}

\item{topN}{Minimum number of terms for each group (default: 50).}

\item{savefile}{Logical indicating whether to save the results to a file (default: FALSE).}

\item{re.folder}{Path to save additional results, such as interactive plots. Default is \code{'./'}.}

\item{enrich_plots}{A logical indicating whether to generate interactive enrichment plots. Default is \code{FALSE}.}
}
\value{
A data.table containing pathway analysis results for all contrasts.
}
\description{
This function conducts pathway analysis using the gprofiler2 package for each contrast defined in the "Contrast" column of the input DMRs table.
The results for each contrast are combined into a single data.table.
}
\references{
For gprofiler2 package documentation, see: https://cran.r-project.org/package=gprofiler2
}
\seealso{
\code{\link{path_results}} for additional details on result formatting.
}
\keyword{DMRs}
\keyword{data.table}
\keyword{enrichment}
\keyword{gprofiler2}
\keyword{methylation}
\keyword{pathway}
