% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_res.R
\name{pca_res}
\alias{pca_res}
\title{Perform Principal Component Analysis (PCA) on Beta Values}
\usage{
pca_res(
  beta_top100,
  ss,
  sampGroups,
  path = "./",
  scale = TRUE,
  center = TRUE,
  filename = "pca_plot_"
)
}
\arguments{
\item{beta_top100}{A matrix of beta values. The rows represent probes, and the columns represent samples.}

\item{ss}{A data frame or data.table containing the sample sheet information.}

\item{sampGroups}{The name of the column in \code{ss} that contains sample group identifiers.}

\item{path}{A character string specifying the directory where plots will be saved (default: "./").}

\item{scale}{Logical, indicating whether to scale the variables (default: TRUE).}

\item{center}{Logical, indicating whether to center the variables (default: TRUE).}

\item{filename}{A name for the saved plot filenames (default: "pca_plot_")}
}
\value{
A \code{prcomp} object representing the results of PCA.
}
\description{
This function applies PCA to a matrix of beta values, visualizes the first four principal
components (PC1 vs. PC2 and PC3 vs. PC4), and saves the resulting plots.
}
\examples{
data("top_500_beta")
data("samplesheet")
pca_res(top_500_beta,samplesheet,sampGroups='Condition',scale = TRUE, center = TRUE,filename='pca_plot')

}
