% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDMP.R
\name{plotDMP}
\alias{plotDMP}
\title{Plot Differentially Methylated Probes (DMPs)}
\usage{
plotDMP(
  DMPann,
  names = c("DMP_count.png", "DMP_count_facet.png", "DMP_annCGI.png",
    "DMP_annGenomic.png"),
  path = NULL
)
}
\arguments{
\item{DMPann}{A data.table containing DMP annotations.}

\item{names}{A character vector specifying the names for saving the plots.}

\item{path}{Path to save the plots. If NULL, plots won't be saved.}
}
\value{
A list of ggplot2 objects representing the generated plots.
}
\description{
This function generates various plots to visualize different aspects of differentially methylated probes (DMPs).
}
\examples{
\dontrun{
# Example usage:
DMP_annotations <- data.table(
  Contrast = c("Group1", "Group1", "Group2", "Group2"),
  Type = c("Hyper", "Hyper", "Hypo", "Hypo"),
  Relation_to_Island = c("Island", "Shore", "Island", "Shore"),
  UCSC_RefGene_Group = c("Gene1;Gene2", "Gene3", "Gene4", ""),
  UCSC_RefGene_Group_short = c("Gene1", "Gene3", "Gene4", ".")
)
plot_list <- plotDMP(
DMPann = DMP_annotations,
names = "DMP_plots",
path = "analysis/DNA_methylation/"
)
}

}
