% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_methods.R
\name{pq}
\alias{pq}
\title{Apply Quantile normalization to RGChannelSet.}
\usage{
pq(rgSet)
}
\arguments{
\item{rgSet}{RGChannelSet object to be normalized.}
}
\value{
Normalized RGChannelSet object.
}
\description{
Apply Quantile normalization to RGChannelSet.
}
