% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{prep}
\alias{prep}
\title{Preprocess Methylation Data}
\usage{
prep(
  mSetSqn,
  remove_sex = TRUE,
  sexchr = c("chrX", "chrY"),
  arraytype = NULL,
  sexplot_folder = NULL,
  predict_sex = TRUE
)
}
\arguments{
\item{mSetSqn}{MethylSet object to be preprocessed.}

\item{remove_sex}{Logical, whether to remove sex chromosomes. Default is TRUE.}

\item{sexchr}{Vector of chromosome names representing sex chromosomes (default is c("chrX", "chrY")).}

\item{arraytype}{Type of array used for preprocessing. If provided, array-specific annotation is used.}

\item{sexplot_folder}{Path to save the sex plot}

\item{predict_sex}{Logical, whether to predict and plot sex. Default is TRUE.}
}
\value{
Preprocessed MethylSet object with predictedSex added in the colData()
}
\description{
This function performs preprocessing steps on a MethylSet object, such as removing probes with known SNPs,
removing cross-reactive probes, predicting and plotting sex, and optionally removing sex chromosomes.
}
\examples{
data("mSet_normalized")
prep(mSet_normalized,remove_sex = TRUE,arraytype='EPICv2',predict_sex = TRUE,sexplot_folder = "./")


}
