% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purify.R
\name{purify}
\alias{purify}
\title{Purify RGChannelSet}
\usage{
purify(rgSet, arraytype = NULL)
}
\arguments{
\item{rgSet}{RGChannelSet object to be purified.}

\item{arraytype}{Array type, specifying the type of Illumina array.}
}
\value{
Purified RGChannelSet.
}
\description{
This function purifies an RGChannelSet by imputing missing values and applying random forest-based correction.
}
\examples{
\dontrun{
# Purify RGChannelSet for EPICv2 array
purify(rgSet, arraytype = "EPICv2")

# Purify RGChannelSet for EPIC or 450K array
purify(rgSet, arraytype = "EPIC")
}

}
