% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC_plots.R, R/QC_plots_factory.R
\name{qc}
\alias{qc}
\title{generate qc plots for signal distribution prior to filtering}
\usage{
qc(
  rgSet,
  sampGroups = NULL,
  sampNames = "Sample_Name",
  cols = NULL,
  qc_folder = "analysis/intermediate/QC/"
)

qc(
  rgSet,
  sampGroups = NULL,
  sampNames = "Sample_Name",
  cols = NULL,
  qc_folder = "analysis/intermediate/QC/"
)
}
\arguments{
\item{rgSet}{rgSet object containing channel intenisty values}

\item{sampGroups}{variables to use for coloring groups}

\item{sampNames}{variable containing barcodes or ids matching colnames of the rgsetdata}

\item{qc_folder}{path to the folder where plots will be saved}

\item{pal}{Color palette to use}
}
\value{
plots

plots
}
\description{
Generate qc plots

Generate qc plots
}
\author{
izar de Villasante
}
