% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_methods.R
\name{quantile}
\alias{quantile}
\title{Apply Quantile normalization to RGChannelSetObject}
\usage{
quantile(rgSet)
}
\arguments{
\item{rgSet}{RGChannelSet object to normalize.}
}
\value{
Normalized RGChannelSet object.
}
\description{
Apply Quantile normalization to RGChannelSetObject
}
