% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_file.R
\name{read_file}
\alias{read_file}
\title{Read File as RDS, Using fread, or readxl for Excel files}
\usage{
read_file(file_path)
}
\arguments{
\item{file_path}{The path to the file to be read.}
}
\value{
Data read from the file if successful.
}
\description{
This function attempts to open a file, first checking if it is an RDS file. If not,
it tries to open the file using \code{data.table::fread()} for CSV files or \code{readxl::read_excel()} for Excel files.
}
