% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_idats_factory.R
\name{read_idats}
\alias{read_idats}
\title{Read and Annotate IDAT Files}
\usage{
read_idats(ss, arraytype = NULL)
}
\arguments{
\item{ss}{A data.table containing the sample sheet file.}

\item{arraytype}{Which platform has been used for the array c("450K","EPIC","EPICv2")}
}
\value{
A list of target objects.
}
\description{
Loads IDAT files into an rgSet, names them correctly, and annotates them based on the array type.
}
