% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_meth.R
\name{read_meth}
\alias{read_meth}
\title{Read IDAT Files and Assign Array Type}
\usage{
read_meth(targets, arraytype = NULL)
}
\arguments{
\item{targets}{A data frame or data.table containing sample metadata, including the paths to the IDAT files.}

\item{arraytype}{A character string specifying the array platform. Options include "450K", "EPIC", "EPICv2", "27K", or "mammal". If \code{NULL} or "guess", the array type will be inferred from the data.}
}
\value{
RGSet object with assigned array type.
}
\description{
Reads IDAT files and assigns the array type based on the provided information or guessed from the data.
}
