% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_plot.R
\name{save_plot}
\alias{save_plot}
\title{Save Plot to File}
\usage{
save_plot(object, filename, width = 480, height = 620, path = "./")
}
\arguments{
\item{object}{A plot object (base R plot or ggplot2 plot).}

\item{filename}{Name of the file to save the plot.}

\item{width}{Width of the plot in inches (default: 480).}

\item{height}{Height of the plot in inches (default: 620).}

\item{path}{Directory path for saving the plot (default: "./").}
}
\value{
The plot object.
}
\description{
This function saves a plot to a file in PNG format.
}
