% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savecoldata.R
\name{savecoldata}
\alias{savecoldata}
\title{Save Column Data from a SummarizedExperiment Object}
\usage{
savecoldata(
  rgset,
  file = "results/ss_clean.csv",
  quote = FALSE,
  check_assertion = TRUE,
  sep = ",",
  dir = ".",
  verbose = FALSE
)
}
\arguments{
\item{rgset}{A SummarizedExperiment object.}

\item{file}{The name of the file to save the data (default: "results/ss_clean.csv").}

\item{quote}{A logical value indicating whether to quote the output file (default: FALSE).}

\item{check_assertion}{Logical, perform an assertion check (default: TRUE).}

\item{sep}{The field separator for the output file (default: ",").}

\item{dir}{The directory where the file should be saved (default: ".").}

\item{verbose}{Logical, print informative messages (default: FALSE).}
}
\value{
A data frame containing the cleaned column data.
}
\description{
This function extracts the column data from a SummarizedExperiment object,
drops unused levels from factors in the resulting data frame, and saves the
data frame to a CSV file. Optionally, it performs an assertion check to ensure
consistency between row names of the SummarizedExperiment object and column
names of the data frame.
}
\examples{
\dontrun{
data("mSet_normalized")
savecoldata(mSet_normalized)
}

}
