% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_methods.R
\name{ssnoob}
\alias{ssnoob}
\title{Apply ssNoob Normalization to an RGChannelSet Object}
\usage{
ssnoob(rgSet)
}
\arguments{
\item{rgSet}{RGChannelSet object to normalize.}
}
\value{
Normalized RGChannelSet object.
}
\description{
This function applies the ssNoob (Single Sample Negative Control Probe) normalization method to an RGChannelSet object.
The ssNoob method is a variant of Noob normalization, designed to address dye bias by using a single sample reference.
}
