% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarize}
\alias{summarize}
\title{Summarize DMPs and DMRs}
\usage{
summarize(dmps, dmrs, path = "results/")
}
\arguments{
\item{dmps}{Data.table containing DMP information.}

\item{dmrs}{Data.table containing DMR information.}

\item{path}{Path to save the summary, defaults to "results/".}
}
\value{
A data.table summarizing DMPs and DMRs.
}
\description{
This function summarizes Differentially Methylated Positions (DMPs) and Differentially Methylated Regions (DMRs).
}
