% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_filter_dmps.R
\name{summary_dmps}
\alias{summary_dmps}
\title{Summarize DMPs}
\usage{
summary_dmps(
  DMPextr_object,
  dir = "./results/dmps/",
  name = "raw",
  write = FALSE
)
}
\arguments{
\item{DMPextr_object}{An object containing DMPs, typically a data table.}

\item{dir}{A character string specifying the directory to save the results (default: "./results/dmps/").}

\item{name}{A character string specifying the name for the summary, used in file naming (default: "raw").}

\item{write}{Logical indicating whether to write the raw DMPs and summary to files (default: FALSE).}
}
\value{
A data table containing the summarized DMPs grouped by contrast.
}
\description{
Generates a summary of DMPs, including total, hypermethylated, and hypomethylated DMPs and their associated genes, grouped by contrast. Optionally writes the raw DMPs and summary to files.
}
\details{
Summarize Differentially Methylated Positions (DMPs)
}
