% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_filter_dmrs.R
\name{summary_dmrs}
\alias{summary_dmrs}
\title{Summarize DMRs}
\usage{
summary_dmrs(dmrs, path = "/results/dmrs/", write = TRUE)
}
\arguments{
\item{dmrs}{Data.table containing DMR information.}

\item{path}{Path to save the summary, defaults to "/results/dmrs/".}

\item{write}{Logical, indicating whether to write the summary to a file.}
}
\value{
A data.table summarizing DMRs by contrast.
}
\description{
This function summarizes Differentially Methylated Regions (DMRs) by contrast.
}
\author{
Izar de Villasante
}
