% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_methods.R
\name{swan}
\alias{swan}
\title{Apply SWAN normalization to RGChannelSet.}
\usage{
swan(rgSet)
}
\arguments{
\item{rgSet}{RGChannelSet object to be normalized.}
}
\value{
Normalized RGChannelSet object.
}
\description{
Apply SWAN normalization to RGChannelSet.
}
