% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanoplot.R
\name{volcanoplot}
\alias{volcanoplot}
\title{Volcano plot}
\usage{
volcanoplot(object, path = "./")
}
\arguments{
\item{object}{A data frame containing differential methylation or region data.
It should include columns \code{logFC}, \code{adj.P.Val}, and \code{Contrast}.}

\item{path}{A string indicating the directory path to save the generated plots. Default is the current directory.}
}
\value{
A volcano plot is saved as a PNG image in the specified path.
}
\description{
This function creates a volcano plot from differential methylation or
region data (DMPs or DMRs), including annotations of the top regulated genes.
}
