% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution.R
\name{celldeconvolution}
\alias{celldeconvolution}
\title{Cell type Deconvolution}
\usage{
celldeconvolution(raw_object, object, arraytype)
}
\arguments{
\item{raw_object}{The input raw RGChannelSet.}

\item{object}{MehtylSet object with normalized data (clean object in the pipeline).}

\item{arraytype}{Vector of the name of the platform. Only supports 'HM450' and 'EPIC'}
}
\value{
A methylset object containing deconvolution data in the colData slot
}
\description{
This function performs deconvolution based on DNA methylation data.
}
