% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpca.R
\name{corpca}
\alias{corpca}
\title{Perform Correlation-based PCA and Generate Plot}
\usage{
corpca(
  beta_top100,
  metadata,
  vars = NULL,
  path = "./",
  filename = "pca_corrplot",
  title = "clinical correlations"
)
}
\arguments{
\item{beta_top100}{A matrix of beta values (rows are probes, columns are samples).}

\item{metadata}{A data frame containing sample metadata.}

\item{vars}{A vector of metadata variable names to include in the eigencorplot. If \code{NULL}, all variables with more than one unique value will be included (default: \code{NULL}).}

\item{path}{Directory path for saving the plot (default: "./").}

\item{filename}{Name of the file to save the plot (default: "pca_corrplot").}

\item{title}{Title for the eigencorplot (default: 'clinical correlations').}
}
\value{
A PCAtools eigencorplot object.
}
\description{
This function performs correlation-based Principal Component Analysis (PCA) on beta values
and generates an eigencorplot to visualize correlations between clinical variables and the principal components.
}
\examples{
\dontrun{
data("top_500_beta")
data("samplesheet")
corpca(top_500_beta,samplesheet)
}

}
