% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_list.R
\name{correlation_analysis}
\alias{correlation_analysis}
\title{Correlation Analysis}
\usage{
correlation_analysis(clean_object, path, variables, sg)
}
\arguments{
\item{clean_object}{A Rgset object.}

\item{path}{The path to save plots.}

\item{variables}{A vector of variable names to include in the analysis, or "ALL" to include all variables.}

\item{sg}{A grouping variable to color the correlation plot.}
}
\value{
A list of correlated variable pairs.
}
\description{
This function performs correlation analysis on a Rgset object. It identifies correlated variables, and generates plots.
}
