% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_list.R
\name{correlation_analysis}
\alias{correlation_analysis}
\title{Correlation analysis}
\usage{
correlation_analysis(clean_object, path, variables, sg)
}
\arguments{
\item{clean_object}{Rgset object}

\item{path}{Path to save the correlation plot}
}
\description{
Correlation analysis
}
