% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denplot.R
\name{denplot}
\alias{denplot}
\title{Density plot}
\usage{
denplot(object, metadata, sampGroups, path = "./", norm_method)
}
\arguments{
\item{object}{An object containing intensity values (not betas).}

\item{metadata}{A data frame containing sample metadata.}

\item{sampGroups}{A character string specifying the column name in \code{metadata} that contains the sample grouping.}

\item{path}{A character string specifying the output directory for the density plot image. Defaults to "./".}

\item{norm_method}{A character string specifying the normalization method used.}
}
\value{
Generates a PNG file with the density plot in the specified path.
}
\description{
This function calculates beta values and generates a density plot.
}
