% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denplot.R
\name{denplot}
\alias{denplot}
\title{Density plot}
\usage{
denplot(object, metadata, sampGroups, path = "./", norm_method)
}
\arguments{
\item{object}{An object containing normalized values (not betas, they are computed internally).}

\item{metadata}{A data frame containing sample metadata.}

\item{sampGroups}{A character string specifying the column name in \code{metadata} that contains the sample grouping.}

\item{path}{A character string specifying the output directory for the density plot image. Defaults to "./".}

\item{norm_method}{A character string specifying the normalization method used for normalization.}
}
\value{
Generates a PNG file with the density plot in the specified path.
}
\description{
This function generates a density plot.
}
\examples{
data(samplesheet)
data(mSet_normalized)
denplot(mSet_normalized,samplesheet,sampGroups='Type',norm_method='swan')

}
