% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_plots.R
\name{distribution_plots}
\alias{distribution_plots}
\title{Distribution plots}
\usage{
distribution_plots(data, variable_interest, path = "./")
}
\arguments{
\item{data}{Data frame containing metadata}

\item{variable_interest}{Name of the variable in the metadata that refers to the sample groups or condition.}

\item{path}{Path to save the plots (default: "./")}
}
\description{
This function generates either boxplots or barplots for all variables in a data frame,
colored by the variable of interest (such as sample groups or conditions).
}
\examples{
data <- data.frame(
  Group = factor(c("A", "B", "A", "B", "A", "B")),
  Age = c(30, 25, 40, 35, 28, 45),
  Gender = factor(c("M", "F", "M", "F", "M", "F"))
)
distribution_plots(data, variable_interest = "Group")
}
