% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtable.R
\name{dtable}
\alias{dtable}
\title{Dynamic table with download buttons}
\arguments{
\item{data}{A data frame to be rendered in the DataTable.}
}
\value{
A DataTable with formatted data and handlers for CSV and Excel downloads.
}
\description{
Render DataTable and handle download events for internal use.
}
\details{
This function renders a DataTable and handles download events for CSV and Excel formats.
It is intended for internal use within the package and should not be exported to the end user.
}
\examples{
\dontrun{
dtable(my_data)
}
}
\keyword{internal}
