% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_filter_dmps.R
\name{filter_dmps}
\alias{filter_dmps}
\title{Filter DMPs Function}
\usage{
filter_dmps(dmps, p.value = 0.01, mDiff = 0.3, s = F)
}
\arguments{
\item{dmps}{Data table containing DMPs.}

\item{p.value}{P-value threshold.}

\item{mDiff}{Mean methylation difference threshold.}

\item{s}{Logical indicating whether to include summary statistics.}
}
\value{
Filtered DMPs.
}
\description{
Filter DMPs Function
}
