% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_filter_dmrs.R
\name{filter_dmrs}
\alias{filter_dmrs}
\title{Filter DMRs}
\usage{
filter_dmrs(dmrs, p.value = 0.05, mDiff = 0.05, min.cpg = 3, s = FALSE)
}
\arguments{
\item{dmrs}{A data table containing DMR information.}

\item{p.value}{Threshold for the False Discovery Rate (default: 0.05).}

\item{mDiff}{Threshold for the absolute mean methylation difference (default: 0.05).}

\item{min.cpg}{Minimum number of CpGs in a DMR (default: 3).}

\item{s}{Logical indicating whether to generate a summary of filtered DMRs (default: FALSE).}
}
\value{
A data table containing the filtered DMRs, or a summary if \code{s = TRUE}.
}
\description{
Filters a data table of DMRs based on user-specified thresholds for False Discovery Rate (FDR), methylation difference, and the minimum number of CpGs. Optionally, a summary of the filtered DMRs is generated.
}
\details{
Filter DMRs
}
