% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_methods.R
\name{funn}
\alias{funn}
\title{Apply FunNorm normalization to RGChannelSet.}
\usage{
funn(rgSet)
}
\arguments{
\item{rgSet}{RGChannelSet object to be normalized.}
}
\value{
Normalized RGChannelSet object.
}
\description{
Apply FunNorm normalization to RGChannelSet.
}
