% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pathways.R
\name{get_pathways}
\alias{get_pathways}
\title{Get Pathways}
\usage{
get_pathways(
  dmrs,
  res.folder = "results/pathways/",
  cols = c("term_size", "query_size", "intersection_size"),
  pval = 0.05,
  topN = 50,
  savefile = FALSE
)
}
\arguments{
\item{dmrs}{Data table containing DMRs.}

\item{res.folder}{Path to save the plots and intermediate files.}

\item{cols}{Columns to include in the results.}

\item{pval}{P-value threshold.}

\item{topN}{Number of pathways to show on the summary.}

\item{savefile}{Logical, whether to save the results to a file.}
}
\value{
Pathway results as a data.table.
}
\description{
Get pathway results based on DMRS
}
\details{
Get Pathways function
}
