% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_top.R
\name{heatmap_top}
\alias{heatmap_top}
\title{Heatmap for beta values}
\usage{
heatmap_top(
  beta_values_matrix,
  metadata,
  sampGroups,
  pal = RColorBrewer::brewer.pal(8, "Dark2"),
  path = "./",
  filename = "heatmap"
)
}
\arguments{
\item{beta_values_matrix}{A matrix of beta values (rows represent CpG probes, columns represent samples).}

\item{metadata}{A data frame containing sample metadata (e.g., sample group information).}

\item{sampGroups}{A string specifying the column name in \code{metadata} that represents the sample group for annotation.}

\item{pal}{Color palette for sample group (default: Brewer palette "Dark2")}

\item{path}{A string indicating the directory path to save the generated heatmap plot. Default is the current directory. (default: "./")}

\item{filename}{Filename of the saved heatmap plot.}
}
\value{
A heatmap object
}
\description{
This function generates a heatmap of CpG probes using beta values.
}
\examples{
data("top_500_beta")
data("samplesheet")
heatmap_top(top_500_beta, samplesheet, sampGroups = "Condition")


}
