% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_top.R
\name{heatmap_top}
\alias{heatmap_top}
\title{Heatmap}
\usage{
heatmap_top(
  top_beta_100,
  ss,
  sampGroups,
  pal = RColorBrewer::brewer.pal(8, "Dark2"),
  path = "./",
  filename = ""
)
}
\arguments{
\item{pal}{Color palette for sample condition(default: Brewer palette "Dark2")}

\item{path}{Path to save the plots (default: "./")}

\item{top_beta_1000}{A matrix of the top 1000 beta values.}

\item{metadata}{A data frame containing sample metadata}
}
\value{
Heatmap
}
\description{
Generate nice HEATMAP for top beta values
}
\details{
This function generates a heatmap of top 1000 variable CpG probes.
}
