% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanoplot.R
\name{manhattanplot}
\alias{manhattanplot}
\title{Manhtattan plot}
\usage{
manhattanplot(object, path = "./")
}
\arguments{
\item{object}{Data frame containing DMPs or DMRs.}
}
\description{
Manhtattan plot
}
