% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.R
\name{mod}
\alias{mod}
\title{Construct Models and Contrasts with limma}
\usage{
mod(
  object,
  betas_idx = NULL,
  group_var = "Sample_Group",
  covs.formula = NULL,
  contrasts = NULL,
  covs = NULL,
  metadata,
  set = TRUE,
  gr = NULL,
  pairwise = TRUE,
  singular = FALSE,
  rename = NULL,
  idcol = "barcode"
)
}
\arguments{
\item{object}{A matrix or data frame containing beta values for the analysis.}

\item{betas_idx}{A matrix or data frame providing the indices for the beta values}

\item{group_var}{A string specifying the column name for the grouping variable.}

\item{covs.formula}{Formula for specifying covariates in the model.}

\item{contrasts}{Optional string of semicolon-separated contrasts. If not provided, automatic contrasts are generated.}

\item{covs}{A character vector specifying additional covariates.}

\item{metadata}{The metadata or sample sheet.}

\item{set}{A boolean vector to subset the observations.}

\item{gr}{A character vector specifying group variables.}

\item{pairwise}{Logical indicating whether to create pairwise group comparisons (default: TRUE).}

\item{singular}{Logical indicating whether to create singular group comparisons (default: FALSE).}

\item{rename}{Character vector to rename the contrasts.}

\item{idcol}{Column name for the sample ID (default: "barcode").}
}
\value{
A limma eBayes model.
}
\description{
This function is used to construct models and contrasts for differential expression analysis using limma.
It builds a design matrix based on the group variable and optional covariates, fits the linear model, and generates contrasts
for pairwise or singular group comparisons.
}
