% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_rgset.R
\name{name_rgset}
\alias{name_rgset}
\title{Rename RGSet and Adjust Column Names}
\usage{
name_rgset(res, targets, idcol = "Sample_Name")
}
\arguments{
\item{res}{An RGSet object to be processed.}

\item{targets}{Target information for column renaming.}

\item{idcol}{Identifier column for matching with the RGSet columns.}
}
\value{
Modified RGSet object with adjusted column names and optional renaming.
}
\description{
This function renames an RGSet object and adjusts column names based on target information.
}
