% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_methods.R
\name{noob}
\alias{noob}
\title{Apply Noob normalization to RGChannelSet.}
\usage{
noob(rgSet)
}
\arguments{
\item{rgSet}{RGChannelSet object to be normalized.}
}
\value{
Normalized RGChannelSet object.
}
\description{
Apply Noob normalization to RGChannelSet.
}
