% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_methods.R
\name{noob_swan}
\alias{noob_swan}
\title{Apply Noob followed by SWAN normalization to RGChannelSet.}
\usage{
noob_swan(rgSet)
}
\arguments{
\item{rgSet}{RGChannelSet object to be normalized.}
}
\value{
Normalized RGChannelSet object.
}
\description{
Apply Noob followed by SWAN normalization to RGChannelSet.
}
