% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_methods.R
\name{normalization_all_methods}
\alias{normalization_all_methods}
\title{Normalize Data Using Multiple Methods and Generate Density Plots}
\usage{
normalization_all_methods(
  object,
  metadata,
  sampGroups,
  path = "./",
  norm_method
)
}
\arguments{
\item{object}{A filtered object (e.g., methylation data) to be normalized}

\item{metadata}{@param metadata A data frame containing sample metadata, which includes a column for grouping samples (e.g., treatment or control groups).}

\item{sampGroups}{A character string specifying the column in the metadata that contains the grouping variable for coloring.}

\item{path}{A character string specifying the path where the density plots will be saved. Default is the current working directory ("./").}

\item{norm_method}{A character string specifying the normalization method to exclude from the plot generation. Options include 'noob', 'ssnoob', 'swan', 'funn', 'quantile', 'noob_pq', and 'noob_swan'.}
}
\description{
This function applies various normalization methods to a given dataset and generates density plots
for each method. The plots are saved to the specified directory. The user has choose to exclude a specific
normalization method by specifying it in the \code{norm_method} argument.
}
