% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway.R
\name{path_results}
\alias{path_results}
\title{Generate Results Excel Sheet for Pathway Analysis}
\usage{
path_results(
  pathway,
  topN = 50,
  method = "method",
  pval = 0.05,
  path = "results/pathways.csv",
  cols = NULL,
  savefile = FALSE
)
}
\arguments{
\item{pathway}{A data.table with pathway analysis results compatible with gopath function output values.}

\item{topN}{Numeric value. The minimum number of terms to include for each group.}

\item{method}{A character or a vector of characters specifying the columns to group by. Default is "method".}

\item{pval}{FDR threshold to filter out terms (default: 0.05).}

\item{path}{The path to save the results CSV file (default: "results/pathways.csv").}

\item{cols}{Columns to include in the pathway results (default: NULL).}

\item{savefile}{Logical indicating whether to save the results to a file (default: FALSE).}
}
\value{
A data.table containing filtered pathway analysis results.

A data.table containing filtered pathway analysis results.
}
\description{
Generate Pathway Results in an Excel Sheet
}
\details{
This function filters pathway analysis results for at least the topN pathways for each group and includes all terms with FDR <= FDR. The filtered results are then saved to a CSV file and returned as a data.table.
}
\examples{
# Example usage:
pathway_results <- data.table(
  Contrast = c("Group1", "Group1", "Group2", "Group2"),
  FDR = c(0.01, 0.02, 0.03, 0.04),
  term_size = c(20, 25, 18, 22),
  query_size = c(15, 18, 12, 16),
  intersection_size = c(10, 12, 8, 11),
  TERM = c("Pathway1", "Pathway2", "Pathway3", "Pathway4"),
  method = c("Method1", "Method1", "Method2", "Method2")
)
result <- path_results(pathway_results)

# Example usage:
pathway_results <- data.table(
  Contrast = c("Group1", "Group1", "Group2", "Group2"),
  FDR = c(0.01, 0.02, 0.03, 0.04),
  term_size = c(20, 25, 18, 22),
  query_size = c(15, 18, 12, 16),
  intersection_size = c(10, 12, 8, 11),
  TERM = c("Pathway1", "Pathway2", "Pathway3", "Pathway4"),
  method = c("Method1", "Method1", "Method2", "Method2")
)
result <- path_results(pathway_results)

}
\seealso{
\code{\link{pathway}} for pathway analysis.

\code{\link{pathway}} for pathway analysis.
}
\author{
izar de Villasante
}
\keyword{analysis}
\keyword{data.table}
\keyword{filtering}
\keyword{pathway}
\keyword{results}
