% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_res.R
\name{pca_res}
\alias{pca_res}
\title{Perform Principal Component Analysis (PCA) on Beta Values}
\usage{
pca_res(
  beta_top100,
  ss,
  sampGroups,
  path = "./",
  scale = TRUE,
  center = TRUE,
  filename = ""
)
}
\arguments{
\item{beta_top100}{A matrix of beta values.}

\item{scale}{Logical, indicating whether to scale the variables (default: TRUE).}

\item{center}{Logical, indicating whether to center the variables (default: TRUE).}
}
\value{
A prcomp object representing the results of PCA.
}
\description{
This function applies Principal Component Analysis (PCA) to a matrix of beta values.
}
\examples{
# Example usage:
# pca_res(beta_top100, scale = TRUE, center = TRUE)

}
