% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_list.R
\name{preparation}
\alias{preparation}
\title{Sample sheet check and preparation}
\usage{
preparation(object)
}
\arguments{
\item{object}{RgSet object that includes metadata}
}
\value{
A data.table containing the processed and relevant variables for correlation analysis.
}
\description{
Check and prepare the sample sheet for correlation analysis by filtering and processing relevant metadata variables.
}
