% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC_plots.R
\name{qc}
\alias{qc}
\title{Generate QC Plots}
\usage{
qc(
  rgSet,
  sampGroups = NULL,
  idcol = "Sample_Name",
  pal = NULL,
  qc_folder = "analysis/intermediate/QC/"
)
}
\arguments{
\item{rgSet}{rgSet object containing channel intenisty values.}

\item{sampGroups}{A variable to use for coloring groups in the plots. Must match a column name in \code{rgSet@colData}.}

\item{idcol}{A column name in \code{rgSet@colData} used as an identifier. Defaults to "Sample_Name".}

\item{pal}{A color palette to use for group coloring. Defaults to \code{NULL} (uses the default palette).}

\item{qc_folder}{A character string specifying the path to the folder where plots will be saved.}
}
\value{
The function saves QC plots as image files in the specified folder.
}
\description{
This function generates various quality control (QC) plots for signal distribution to assess the quality of the data before normalization.
}
\details{
This function generates:
\itemize{
\item A QC report in PDF format using \code{minfi::qcReport}.
\item A density plot showing beta value distributions before normalization.
\item A mean QC plot showing the mean signal intensities.
}
}
\examples{

data("rgSet")

qc(rgSet,
sampGroups = 'Type',
idcol='Sample_Name',
qc_folder = './')


}
