% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_files.R
\name{read_file}
\alias{read_file}
\title{Read files}
\usage{
read_file(file_path)
}
\arguments{
\item{file_path}{A string specifying the path to the file  (include it. eg. 'data/file.csv').}
}
\value{
A data frame or R object, depending on the file type.
}
\description{
Read files
}
