% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_idats.R
\name{read_idats}
\alias{read_idats}
\title{Read and Annotate IDAT Files}
\usage{
read_idats(
  ss,
  arraytype = NULL,
  idats_folder = NULL,
  idcol = NULL,
  description = "",
  author = "ijcBIT"
)
}
\arguments{
\item{ss}{A data.table containing the sample sheet file.}

\item{arraytype}{A character string specifying the array platform used. (options: "450K", "EPIC", "EPICv2"). If \code{NULL}, the array type will be inferred from the sample sheet if available.}

\item{idats_folder}{Optional. A character string specifying the path to the folder containing IDAT files, if it is not correctly specified in the \code{Basename} column of the samplesheet.}

\item{idcol}{The column name in the sample sheet that contains unique identifiers for samples.(eg. Sample_Name)}

\item{description}{A description for the rgSet (default: "").}

\item{author}{The author of the rgSet (default: "ijcBIT").}
}
\value{
A SummarizedExperiment object.
}
\description{
Load IDAT files into a rgSet, names them correctly, and annotates them based on the array type.
}
\details{
The "Basename" column contains the path to the idat files (without _Red.idat and Grn.idat)
If idats_folder is provided, it will generate the "Basename" column by combining the \code{idats_folder}, \code{Sentrix_ID},
and \code{Sentrix_Position} columns.
If idats_folder is not provided, the function reads the methylation data with the path that is in the \code{Basename} column.
}
\examples{
\dontrun{

# EXAMPLE 1: Basename column in the samplesheet
# Load the sample sheet
data("samplesheet")

rgSet <- read_idats(ss = samplesheet,
                    arraytype = 'EPICv2',
                    idats_folder = NULL,
                    idcol = "Sample_Name")


# EXAMPLE 2: Idats_folder is provided

data("samplesheet")

# Define the path to the folder containing the IDAT files
idats_folder <-system.file("extdata/EPICv2/idats/", package = "epipe")

rgSet <- read_idats(ss = samplesheet,
                    arraytype = 'EPICv2',
                    idats_folder = idats_folder,
                    idcol = "Sample_Name")
}
}
