% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_idats.R
\name{read_idats}
\alias{read_idats}
\title{Read and Annotate IDAT Files}
\usage{
read_idats(
  ss,
  arraytype = NULL,
  idats_folder = NULL,
  idcol = NULL,
  description = "",
  author = "ijcBIT"
)
}
\arguments{
\item{ss}{A data.table containing the sample sheet file.}

\item{arraytype}{A character string specifying the array platform used. (options: "450K", "EPIC", "EPICv2"). If \code{NULL}, the array type will be inferred from the sample sheet if available.}

\item{idats_folder}{A character string specifying the path to the folder containing IDAT files.}

\item{idcol}{The column name in the sample sheet that contains unique identifiers for samples.}

\item{description}{A description for the rgSet (default: "").}

\item{author}{The author of the rgSet (default: "ijcBIT").}
}
\value{
A SummarizedExperiment object.
}
\description{
Loads IDAT files into an rgSet, names them correctly, and annotates them based on the array type.
}
\details{
The function will automatically generate the "Basename" column in the sample sheet if it does not already exist,
by combining the \code{idats_folder}, \code{Sentrix_ID}, and \code{Sentrix_Position} columns. It then reads the methylation data and annotates the resulting \code{rgSet} with metadata such as description, author, and category.
}
