% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_idats_factory.R, R/read_meth.R
\name{read_meth}
\alias{read_meth}
\title{Read IDAT Files and Assign Array Type}
\usage{
read_meth(targets, arraytype = NULL)

read_meth(targets, arraytype = NULL)
}
\arguments{
\item{arraytype}{Which platform has been used for the array c("450K","EPIC","EPICv2")}
}
\value{
RGSet object with assigned array type.

RGSet object with assigned array type.
}
\description{
Reads IDAT files and assigns the array type based on the provided information or guessed from the data.

Reads IDAT files and assigns the array type based on the provided information or guessed from the data.
}
