% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_methods.R
\name{ssnoob}
\alias{ssnoob}
\title{Apply ssNoob normalization to RGChannelSet.}
\usage{
ssnoob(rgSet)
}
\arguments{
\item{rgSet}{RGChannelSet object to be normalized.}
}
\value{
Normalized RGChannelSet object.
}
\description{
Apply ssNoob normalization to RGChannelSet.
}
