% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_filter_dmps.R
\name{summary_dmps}
\alias{summary_dmps}
\title{Summarize DMPs and Write Results to File}
\usage{
summary_dmps(
  DMPextr_object,
  dir = "./results/dmps/",
  name = "raw",
  write = FALSE
)
}
\arguments{
\item{DMPextr_object}{Object containing DMPs.}

\item{dir}{Directory to save the results.}

\item{name}{Name for the summary (used in file naming).}

\item{write}{Logical indicating whether to write results to file (default is FALSE).}
}
\value{
Summary of DMPs.
}
\description{
Summarize DMPs and Write Results to File
}
