% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_filter_dmrs.R
\name{summary_dmrs}
\alias{summary_dmrs}
\title{Summarize DMRs}
\usage{
summary_dmrs(dmrs, contrast = NULL, path = "/results/dmrs/", write = TRUE)
}
\arguments{
\item{dmrs}{A data table containing DMRs, including information on \code{Contrast}, \code{meandiff}, and \code{overlapping.genes}.}

\item{contrast}{Optional, a specific contrast to summarize (default: NULL).}

\item{path}{Path to save the summary file (default: "/results/dmrs/").}

\item{write}{Logical, whether to save the summary to file (default: TRUE).}
}
\value{
A data table summarizing DMRs with counts of Hyper and Hypo DMRs, and associated genes.
}
\description{
Summarizes the information of DMRs, calculating the total number of DMRs, Hyper and Hypo DMRs, and the associated genes for each contrast. Optionally, writes the summary to a specified file.
}
