% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot_deconvolution.R
\name{ttest_boxplot}
\alias{ttest_boxplot}
\title{Boxplot and statistical test}
\usage{
ttest_boxplot(data, sampGroup, path = "./", filename = "Cell_type_prop.png")
}
\arguments{
\item{data}{A data frame containing metadata and cell type proportions.}

\item{sampGroup}{A character string specifying the name of the variable in the metadata that refers to the sample groups or condition.}

\item{path}{A character string specifying the directory to save the plot (default: "./").}

\item{filename}{A character string specifying the file name for the saved plot (default: "Cell_type_prop.png").}
}
\value{
Saves a boxplot as a PNG file.
}
\description{
This function performs a t-test to check for statistical differences in cell type proportions between two groups
and creates a boxplot with statistical annotations.
}
\examples{
data("ss_all_variables")
ttest_boxplot(ss_all_variables,sampGroup = 'Condition')
}
