% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot_deconvolution.R
\name{ttest_boxplot}
\alias{ttest_boxplot}
\title{ttest_boxplot}
\usage{
ttest_boxplot(data, sampGroup, path = "./", filename = "Cell_type_prop.png")
}
\arguments{
\item{data}{Data frame containing metadata}

\item{sampGroup}{Name of the variable in the metadata that refers to the sample Grops or condition.}

\item{path}{Path to save the plots (default: "./")}
}
\description{
Boxplot and statistical test
}
\details{
Performs t.test to check for statistical differences for cell type deconvolution between two groups and creates a boxplot.
}
